+++ SceneRender/P11_Fragments2FB.cpp
@@ -18,18 +18,18 @@ void P11_Fragments2FB::processTriangles(Context *context)
     for (ListNode<Fragment> *fptr = ptr->fragments.head; fptr; fptr = fptr->next)
     {
       // get this frament's z-coordinate
-      float z = fptr->zCoord;
+      double z = fptr->zCoord;
       // get this fragment's x and y index into the viewport
       int x = fptr->xIndex;
       int y = fptr->yIndex;
       // get the z-buffer value currently at that pixel of the viewport
-      float zBuff = fb->getDepthVP(x, y);
+      double zBuff = fb->getDepthVP(x, y);
       // if the current fragment is in front of what is in the framebuffer,
       // then overwrite the data in the framebuffer
       if (z < zBuff)
       {
         //unsigned char c[3];
-        //c[0] = (fptr->r); // convert from float to unsigned char
+        //c[0] = (fptr->r); // convert from double to unsigned char
         //c[1] = (fptr->g);
         //c[2] = (fptr->b); /* not yet using the alpha channel */
         fb->setPixelVP(x, y, fptr->pixel, z);
